backend/requirements.txt
playwright==1.58.0
boto3==1.42.73
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
python-dotenv==1.2.2
pypdf==5.6.0
httpx==0.28.1
apscheduler==3.10.4
fastapi
uvicorn[standard]
google-auth-oauthlib
google-api-python-client
python-jose[cryptography]